refactor configs; removed unused vals from EtlConfig case class
